requirements.txt
altgraph==0.17.4
importlib-metadata==7.0.1
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyodbc==5.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pywin32==306
pywin32-ctypes==0.2.2
requests==2.32.3
Wand==0.6.13
zipp==3.17.0
